requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.23.1
transformers>=4.0.0
torch>=2.0.0
scikit-learn>=1.2.2
pydantic>=2.1.1
python-dotenv>=1.0.0
